+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -93,5 +93,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-12-12T13:23:17.402Z")
+    updated: Optional[datetime] = parse_datetime("2025-12-15T06:07:20.045Z")
     r"""The date and time the resource was last updated"""