+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -239,7 +239,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KC9BP7W794VYM74NZJDZ1ZDT"
+        "01KCG9Y25S8ZHM74H4RDG3D7TF"
     )
     r"""The unique identifier of the data source"""