+++ packages/orq-rc/src/orq_ai_sdk/models/listconversationsop.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 from .conversationresponse import ConversationResponse, ConversationResponseTypedDict
 from orq_ai_sdk.types import BaseModel
 from orq_ai_sdk.utils import FieldMetadata, QueryParamMetadata
+import pydantic
 from typing import List, Literal, Optional
 from typing_extensions import Annotated, NotRequired, TypedDict
 
@@ -12,9 +13,9 @@ class ListConversationsRequestTypedDict(TypedDict):
     limit: NotRequired[int]
     r"""A limit on the number of objects to be returned. Limit can range between 1 and 100, and the default is 10."""
     starting_after: NotRequired[str]
-    r"""A cursor for use in pagination. `starting_after` is a conversation ID that defines your place in the list."""
+    r"""A cursor for use in pagination. `startingAfter` is a conversation ID that defines your place in the list."""
     ending_before: NotRequired[str]
-    r"""A cursor for use in pagination. `ending_before` is a conversation ID that defines your place in the list."""
+    r"""A cursor for use in pagination. `endingBefore` is a conversation ID that defines your place in the list."""
 
 
 class ListConversationsRequest(BaseModel):
@@ -26,15 +27,17 @@ class ListConversationsRequest(BaseModel):
 
     starting_after: Annotated[
         Optional[str],
+        pydantic.Field(alias="startingAfter"),
         FieldMetadata(query=QueryParamMetadata(style="form", explode=True)),
     ] = None
-    r"""A cursor for use in pagination. `starting_after` is a conversation ID that defines your place in the list."""
+    r"""A cursor for use in pagination. `startingAfter` is a conversation ID that defines your place in the list."""
 
     ending_before: Annotated[
         Optional[str],
+        pydantic.Field(alias="endingBefore"),
         FieldMetadata(query=QueryParamMetadata(style="form", explode=True)),
     ] = None
-    r"""A cursor for use in pagination. `ending_before` is a conversation ID that defines your place in the list."""
+    r"""A cursor for use in pagination. `endingBefore` is a conversation ID that defines your place in the list."""
 
 
 ListConversationsObject = Literal["list",]