+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KC9BP7W46AFY7ACQK8H0T968"
+        "01KCG9Y25PBRCSAMN2NCYAWSV5"
     )
     r"""The unique identifier of the data source"""