+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KC9BP7W6HBFZY5PVD8W8F2T5"
+        "01KCG9Y25RMSE8BDA93Z5BY4DP"
     )
     r"""The unique identifier of the data source"""