+++ packages/orq-rc/src/orq_ai_sdk/models/updatebudgetop.py
@@ -219,5 +219,5 @@ class UpdateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-12-12T13:23:17.825Z")
+    updated: Optional[datetime] = parse_datetime("2025-12-15T06:07:20.408Z")
     r"""The date and time the resource was last updated"""