+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KC9BP7W9NMHRDHVQGNV3JYTH"
+        "01KCG9Y25VPKK3NR6KYP1JS2BN"
     )
     r"""The unique identifier of the data source"""