+++ packages/orq-rc/RELEASES.md
@@ -4409,3 +4409,13 @@ Based on:
 - [python v4.1.0-rc.41] packages/orq-rc
 ### Releases
 - [PyPI v4.1.0-rc.41] https://pypi.org/project/orq-ai-sdk/4.1.0-rc.41 - packages/orq-rc
+
+## 2025-12-15 06:09:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.676.1 (2.781.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.1.0-rc.42] packages/orq-rc
+### Releases
+- [PyPI v4.1.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.1.0-rc.42 - packages/orq-rc